handle duplicate competitor error more gracefully in competitors copy
"Validointi epäonnistui: Tässä sarjassa on jo saman niminen kilpailija"

If the target race already has a competitor with the same name as
a competitor in the source race, the competitor is skipped. There will still
be an error with duplicate number in case the target competitor number
is such that another competitor in the source race has.